% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkVectorInEqualSizeFragments.R
\name{chunkVectorInEqualSizeFragments}
\alias{chunkVectorInEqualSizeFragments}
\title{Split Vector into Equal-Sized Fragments}
\usage{
chunkVectorInEqualSizeFragments(x, n)
}
\arguments{
\item{x}{A vector to split.}

\item{n}{Integer. The maximum size of each fragment.}
}
\value{
A list of vector fragments.
}
\description{
Splits a vector into fragments of approximately equal size. This is
used for memory-efficient processing when comparing proteomes in chunks.
}
\examples{
chunkVectorInEqualSizeFragments(1:100, n = 20)

}
