% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteProteome.R
\name{deleteProteome}
\alias{deleteProteome}
\title{Delete Proteomes from the Database}
\usage{
deleteProteome(dbConn, proteomeIdVec)
}
\arguments{
\item{dbConn}{A DBI database connection object.}

\item{proteomeIdVec}{Integer vector. Proteome IDs to delete.}
}
\value{
NULL (invisible). The function deletes data from the database.
}
\description{
Deletes one or more proteomes and all associated data using the
\code{taxogeno.delete_proteome} stored procedure.
}
\details{
The deletion is performed via a PostgreSQL stored procedure that
handles cascade deletion of all related records including:
\itemize{
\item GO Slim summaries
\item Distance matrices
\item Gene records
\item All annotation relationships (GO, Keywords, EC, Pathways)
}
}
\examples{
\dontrun{
conn <- dbConnect(RPostgres::Postgres(), dbname = "taxogeno")
deleteProteome(conn, c(1, 2, 3))
}

}
