% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/getGeneratedGoslimNormalizedAnnotationForProteomeIdVec.R
\name{getGeneratedGoslimNormalizedAnnotationForProteomeIdVec}
\alias{getGeneratedGoslimNormalizedAnnotationForProteomeIdVec}
\title{Get Generated GO Slim Normalized Annotation for Proteomes}
\usage{
getGeneratedGoslimNormalizedAnnotationForProteomeIdVec(dbConn, proteomeIdVec)
}
\arguments{
\item{dbConn}{A DBI database connection object.}

\item{proteomeIdVec}{Integer vector. Proteome IDs to retrieve.}
}
\value{
A data frame with columns:
\item{proteomeid}{Proteome identifier}
\item{annotkwid}{GO Slim term ID}
\item{annotkwcount}{Raw annotation count}
\item{annotkwrelval}{Normalized value (count/genecount)}
}
\description{
Retrieves normalized GO Slim annotation vectors for a set of proteomes
from the generated_goslim_summary table.
}
\examples{
\dontrun{
norm_df <- getGeneratedGoslimNormalizedAnnotationForProteomeIdVec(conn, c(1, 2, 3))
}

}
