% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/getKeywordNormalizedAnnotationForProteomeIdVec.R
\name{getKeywordNormalizedAnnotationForProteomeIdVec}
\alias{getKeywordNormalizedAnnotationForProteomeIdVec}
\title{Get Keyword Normalized Annotation for Proteomes}
\usage{
getKeywordNormalizedAnnotationForProteomeIdVec(dbConn, proteomeIdVec)
}
\arguments{
\item{dbConn}{A DBI database connection object.}

\item{proteomeIdVec}{Integer vector. Proteome IDs to retrieve.}
}
\value{
A data frame with columns:
\item{proteomeid}{Proteome identifier}
\item{annotkwid}{Keyword ID}
\item{annotkwrelval}{Normalized value (count/genecount)}
}
\description{
Retrieves normalized keyword annotation vectors for a set of proteomes.
Uses a Cartesian product of proteomes and all available keywords.
}
\examples{
\dontrun{
norm_df <- getKeywordNormalizedAnnotationForProteomeIdVec(conn, c(1, 2, 3))
}

}
