% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastProteomeId.R
\name{getLastProteomeId}
\alias{getLastProteomeId}
\title{Get the Last (Maximum) Proteome ID}
\usage{
getLastProteomeId(dbConn)
}
\arguments{
\item{dbConn}{A DBI database connection object.}
}
\value{
Integer. The maximum proteome ID, or NA if no proteomes exist.
}
\description{
Retrieves the highest proteome ID from the proteome table.
}
\examples{
\dontrun{
conn <- dbConnect(RPostgres::Postgres(), dbname = "taxogeno")
last_id <- getLastProteomeId(conn)
}

}
