% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbiAssemblyInfoFromGcaIdContainingString.R
\name{ncbiAssemblyInfoFromGcaIdContainingString}
\alias{ncbiAssemblyInfoFromGcaIdContainingString}
\title{Retrieve NCBI Assembly Information from GCA ID}
\usage{
ncbiAssemblyInfoFromGcaIdContainingString(dbConn, gcaIdContainingString)
}
\arguments{
\item{dbConn}{A DBI database connection object.}

\item{gcaIdContainingString}{Character string containing a GCA ID
(e.g., "GCA_000001405.28_GRCh38.p14"). The function extracts the GCA ID
using regex.}
}
\value{
A list containing assembly information from the NCBI summary table:
\item{assembly_accession}{GCA accession}
\item{bioproject}{BioProject ID}
\item{biosample}{BioSample ID}
\item{wgs_master}{WGS master accession}
\item{refseq_category}{RefSeq category}
\item{taxid}{NCBI taxonomy ID}
\item{species_taxid}{Species taxonomy ID}
\item{organism_name}{Scientific name}
\item{infraspecific_name}{Strain/isolate name}
\item{isolate}{Isolate identifier}
\item{version_status}{Assembly version status}
\item{assembly_level}{Assembly level (Complete, Chromosome, Scaffold, Contig)}
\item{release_type}{Release type}
\item{genome_rep}{Genome representation}
\item{date}{Release date}
\item{asm_name}{Assembly name}
\item{submitter}{Submitting organization}
\item{gbrs_paired_asm}{GenBank/RefSeq paired assembly}
\item{paired_asm_comp}{Paired assembly comparison}
\item{ftp_path}{FTP path for data download}
\item{excluded_from_refseq}{Excluded from RefSeq flag}
\item{relation_to_type_material}{Type material relationship}
}
\description{
Queries the NCBI assembly summary database for information about an assembly
identified by a GCA accession number.
}
\examples{
\dontrun{
conn <- dbConnect(RPostgres::Postgres(), dbname = "taxogeno")
info <- ncbiAssemblyInfoFromGcaIdContainingString(conn, "GCA_000001405.28")
print(info$organism_name)
}

}
