% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveKeywordEuclideanDistancesForProteomeId.R
\name{saveKeywordEuclideanDistancesForProteomeId}
\alias{saveKeywordEuclideanDistancesForProteomeId}
\title{Save Keyword-Based Euclidean Distances for a Proteome}
\usage{
saveKeywordEuclideanDistancesForProteomeId(dbConn, proteomeId)
}
\arguments{
\item{dbConn}{A DBI database connection object.}

\item{proteomeId}{Integer. The proteome ID to compare.}
}
\value{
NULL (invisible). The function writes distances to
\code{taxogeno.euclidean_distances_keyword}.
}
\description{
Calculates and caches Euclidean distances based on keyword annotations
between a specific proteome and all other proteomes in the database.
}
\details{
This function is a wrapper around the more general
\code{\link{saveEuclideanDistancesForProteomeId}} specifically for
keyword annotations. It performs chunked processing to manage memory.
}
\examples{
\dontrun{
conn <- dbConnect(RPostgres::Postgres(), dbname = "taxogeno")
saveKeywordEuclideanDistancesForProteomeId(conn, 1)
}

}
