% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateBiosqlNcbiTaxonomy.R
\name{updateBiosqlNcbiTaxonomy}
\alias{updateBiosqlNcbiTaxonomy}
\title{Update BioSQL NCBI Taxonomy}
\usage{
updateBiosqlNcbiTaxonomy(dbConn)
}
\arguments{
\item{dbConn}{A DBI database connection object.}
}
\value{
NULL (invisible). The function updates the database directly.
}
\description{
Downloads the latest NCBI taxonomy dump and loads it into the biosql
schema using the \code{load_ncbi_taxonomy.pl} Perl script.
}
\details{
The function:
\enumerate{
\item Downloads \code{taxdump.tar.gz} from NCBI FTP
\item Extracts the archive
\item Executes the Perl script \code{load_ncbi_taxonomy.pl}
\item Cleans up temporary files
}
}
\examples{
\dontrun{
conn <- dbConnect(RPostgres::Postgres(), dbname = "taxogeno")
updateBiosqlNcbiTaxonomy(conn)
}

}
