% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNcbiAssemblyInfoForProteomeIdVec.R
\name{updateNcbiAssemblyInfoForProteomeIdVec}
\alias{updateNcbiAssemblyInfoForProteomeIdVec}
\title{Update NCBI Assembly Information for Proteomes}
\usage{
updateNcbiAssemblyInfoForProteomeIdVec(dbConn, proteomeIdVec)
}
\arguments{
\item{dbConn}{A DBI database connection object.}

\item{proteomeIdVec}{Integer vector. Proteome IDs to update.}
}
\value{
NULL (invisible). The function updates the database directly.
}
\description{
Updates NCBI assembly information (GCA, taxonomy ID, source URL) for
a set of proteomes by querying the NCBI assembly summary database.
}
\details{
For each proteome:
\enumerate{
\item Retrieves existing GCA ID or uses basename as fallback
\item Queries NCBI assembly summary for updated information
\item Updates taxonomy if taxon has changed
\item Inserts new taxon if it doesn't exist
\item Updates proteome record with new assembly info
}
}
\examples{
\dontrun{
conn <- dbConnect(RPostgres::Postgres(), dbname = "taxogeno")
updateNcbiAssemblyInfoForProteomeIdVec(conn, c(1, 2, 3))
}

}
